pyproject.toml
[project]
name = "blobapi"
version = "0.1.0"
description = "OpenAPI spec catalog, normalization, and query generation"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "sqlalchemy>=2.0",
    "httpx",
    "pyyaml",
    "psycopg[binary]>=3.3.3",
    "pyodbc>=5.3.0",
    "duckdb>=1.5.0",
    "duckdb-engine>=0.17.0",
    "dulwich>=1.1.0",
]

[project.optional-dependencies]
sqlite = []  # built-in
duckdb = ["duckdb", "duckdb-engine"]
pg = ["psycopg[binary]"]
sqlserver = ["pyodbc"]
all = ["duckdb", "duckdb-engine", "psycopg[binary]", "pyodbc"]
dev = ["pytest"]
